+++ test_settings.py
@@ -2,8 +2,10 @@ HELPER_SETTINGS = {
     "TIME_ZONE": "America/Chicago",
     "INSTALLED_APPS": [
         "djangocms_alias",
+        'polymorphic',
         "djangocms_content_expiry",
         "djangocms_content_expiry.test_utils.polls",
+        "djangocms_content_expiry.test_utils.polymorphic_project",
         "djangocms_versioning",
         "djangocms_moderation",
         "rangefilter"